@@ -2,10 +2,8 @@ name: nomadlodge_backend_client
 description: Starting point for a Serverpod client.
 
 environment:
-  sdk: '>=3.0.0 <4.0.0'
+  sdk: '>=3.3.0 <4.0.0'
 
 dependencies:
-  serverpod_client: 2.1.1
-  serverpod_auth_client: 2.1.1
-  serverpod_auth_email_flutter: ^2.1.1
-  serverpod_auth_shared_flutter: ^2.1.1
+  serverpod_client: 2.1.4
+  serverpod_auth_client: 2.1.4
