@@ -4,11 +4,12 @@ description: Starting point for a Serverpod server.
 # homepage: https://www.example.com
 
 environment:
-  sdk: '>=3.0.0 <4.0.0'
+  sdk: '>=3.3.0 <4.0.0'
 
 dependencies:
-  serverpod: 2.1.1
-  serverpod_auth_server: 2.1.1
+  serverpod: 2.1.4
+  serverpod_auth_server: 2.1.4
 
 dev_dependencies:
-  lints: ^3.0.0
+  lints: '>=3.0.0 <6.0.0'
+
