@@ -5,8 +5,7 @@ import 'package:nomadlodge_backend_server/src/web/routes/root.dart';
 import 'src/generated/protocol.dart';
 import 'src/generated/endpoints.dart';
 
-import 'package:serverpod_auth_server/module.dart' as auth;
-import 'package:serverpod/protocol.dart';
+import 'package:serverpod_auth_server/serverpod_auth_server.dart' as auth;
 
 // This is the starting point of your Serverpod server. In most cases, you will
 // only need to make additions to this file if you add future calls,  are
@@ -18,6 +17,7 @@ void run(List<String> args) async {
     args,
     Protocol(),
     Endpoints(),
+    authenticationHandler: auth.authenticationHandler,
   );
 
   // If you are using any future calls, they need to be registered here.
@@ -32,12 +32,20 @@ void run(List<String> args) async {
     '/*',
   );
 
-  // Add the authentication module.
-  auth.AuthConfig.set(
-    auth.AuthConfig(
-      firebaseServiceAccountKeyJson: 'config/firebase-service-account.json',
-    ),
-  );
+   auth.AuthConfig.set(auth.AuthConfig(
+    //firebaseServiceAccountKeyJson: 'config/firebase-service-account.json',
+    sendValidationEmail: (session, email, validationCode) async {
+      // TODO: integrate with mail server
+      print('Validation code: $validationCode');
+      return true;
+    },
+    sendPasswordResetEmail: (session, userInfo, validationCode) async {
+      // TODO: integrate with mail server
+      print('Validation code: $validationCode');
+      return true;
+    },
+  ));
+
   // Start the server.
   await pod.start();
 }
