@@ -8,18 +8,102 @@
 // ignore_for_file: type_literal_in_constant_pattern
 // ignore_for_file: use_super_parameters
 
-library protocol;
+library protocol; // ignore_for_file: no_leading_underscores_for_library_prefixes
 
-export 'booking.dart';
-export 'fee.dart';
-export 'geo_address.dart';
-export 'location.dart';
-export 'location_team.dart';
-export 'maintenace.dart';
-export 'maintenance_type.dart';
-export 'media.dart';
-export 'media_type.dart';
-export 'task.dart';
-export 'user.dart';
-export 'user_device.dart';
-export 'user_type.dart';
+import 'package:serverpod/serverpod.dart' as _i1;
+import 'package:serverpod/protocol.dart' as _i2;
+import 'package:serverpod_auth_server/serverpod_auth_server.dart' as _i3;
+import 'example.dart' as _i4;
+export 'example.dart';
+
+class Protocol extends _i1.SerializationManagerServer {
+  Protocol._();
+
+  factory Protocol() => _instance;
+
+  static final Protocol _instance = Protocol._();
+
+  static final List<_i2.TableDefinition> targetTableDefinitions = [
+    ..._i3.Protocol.targetTableDefinitions,
+    ..._i2.Protocol.targetTableDefinitions,
+  ];
+
+  @override
+  T deserialize<T>(
+    dynamic data, [
+    Type? t,
+  ]) {
+    t ??= T;
+    if (t == _i4.Example) {
+      return _i4.Example.fromJson(data) as T;
+    }
+    if (t == _i1.getType<_i4.Example?>()) {
+      return (data != null ? _i4.Example.fromJson(data) : null) as T;
+    }
+    try {
+      return _i3.Protocol().deserialize<T>(data, t);
+    } on _i1.DeserializationTypeNotFoundException catch (_) {}
+    try {
+      return _i2.Protocol().deserialize<T>(data, t);
+    } on _i1.DeserializationTypeNotFoundException catch (_) {}
+    return super.deserialize<T>(data, t);
+  }
+
+  @override
+  String? getClassNameForObject(Object? data) {
+    String? className = super.getClassNameForObject(data);
+    if (className != null) return className;
+    if (data is _i4.Example) {
+      return 'Example';
+    }
+    className = _i2.Protocol().getClassNameForObject(data);
+    if (className != null) {
+      return 'serverpod.$className';
+    }
+    className = _i3.Protocol().getClassNameForObject(data);
+    if (className != null) {
+      return 'serverpod_auth.$className';
+    }
+    return null;
+  }
+
+  @override
+  dynamic deserializeByClassName(Map<String, dynamic> data) {
+    if (data['className'] == 'Example') {
+      return deserialize<_i4.Example>(data['data']);
+    }
+    if (data['className'].startsWith('serverpod.')) {
+      data['className'] = data['className'].substring(10);
+      return _i2.Protocol().deserializeByClassName(data);
+    }
+    if (data['className'].startsWith('serverpod_auth.')) {
+      data['className'] = data['className'].substring(15);
+      return _i3.Protocol().deserializeByClassName(data);
+    }
+    return super.deserializeByClassName(data);
+  }
+
+  @override
+  _i1.Table? getTableForType(Type t) {
+    {
+      var table = _i3.Protocol().getTableForType(t);
+      if (table != null) {
+        return table;
+      }
+    }
+    {
+      var table = _i2.Protocol().getTableForType(t);
+      if (table != null) {
+        return table;
+      }
+    }
+    return null;
+  }
+
+  @override
+  List<_i2.TableDefinition> getTargetTableDefinitions() =>
+      targetTableDefinitions;
+
+  @override
+  String getModuleName() => 'nomadlodge_backend';
+}
